requirements.txt
streamlit
pandas
openpyxl
reportlab
python-docx 